added command line argument 'addseeder' to take the ip address if dns seeder nodes used for node discovery
